refactor: rebuild learning 03，merge by dimension

+++ src/main/java/com/zion/common/vo/learning/response/SubjectVO.java
@@ -28,7 +28,7 @@ public class SubjectVO implements Serializable {
     /**
      * 已延期天数
      */
-    private Integer delayDays;
+    private Long delayDays;
 
 
 }
